feat: restructure Redux files into dedicated redux folder and update imports
