+++ src/lib/kepler/road-roughness-data.ts
@@ -1,4 +1,4 @@
-import store from "@/store";
+import store from "@/lib/redux/store";
 import { addDataToMap } from "@kepler.gl/actions";
 import { processGeojson } from "@kepler.gl/processors";
 import configJson from "./config.json";
@@ -33,3 +33,6 @@ export async function addRoadRoughnessData() {
     }),
   );
 }
+
+
+
